Move broken new mongo config to a separate file to retry old naive one
